requirements.txt
# Crowd Counting Application - Dependencies
# Optimized for CPU inference with minimal footprint

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# PyTorch (CPU-only for deployment)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# Image Processing
pillow==10.2.0
opencv-python-headless==4.9.0.80

# Scientific Computing
numpy==1.26.3
scipy==1.12.0
matplotlib==3.8.2
